# deliveryline-backend/src/main/resources/openapi/openapi.json
{
  "components" : {
    "schemas" : {
...
      "ProblemDetailsResponse" : {
        "description" : "RFC 9457 Problem Details payload with DeliveryLine extension fields.",
        "properties" : {
          "code" : {
            "example" : "RUN_NOT_FOUND",
            "type" : "string"
          },
          "correlationId" : {
            "description" : "Echoed correlation ID when present on the request or MDC.",
            "example" : "3fa85f64-5717-4562-b3fc-2c963f66afa6",
            "type" : [
              "string",
              "null"
            ]
          },
          "detail" : {
            "example" : "Workflow run not found: run_abc123",
            "type" : "string"
          },
          "details" : {
            "anyOf" : [
              {
                "additionalProperties" : { },
                "type" : "object"
              },
              {
                "items" : {
                  "type" : "object"
                },
                "type" : "array"
              }
            ],
            "description" : "Domain-specific extension payload. Validation errors emit an array of field-error objects; other domain errors typically emit an object."
          },
          "instance" : {
            "example" : "/api/v1/workflows/run_abc123/events",
            "type" : "string"
          },
          "retryable" : {
            "example" : false,
            "type" : "boolean"
          },
          "status" : {
            "example" : 404,
            "format" : "int32",
            "type" : "integer"
          },
          "title" : {
            "example" : "Workflow run not found",
            "type" : "string"
          },
          "type" : {
            "example" : "https://deliveryline.local/problems/run-not-found",
            "format" : "uri",
            "type" : "string"
          }
        },
        "required" : [
          "code",
          "detail",
          "instance",
          "retryable",
          "status",
          "title",
          "type"
        ],
        "type" : "object"
      },
...
    }
  },
  "info" : {
    "description" : "Governed local-first agent delivery workflow API. Phase 1 is localhost-bound (loopback-only) and unauthenticated; the embedded server refuses to bind a non-loopback address without an explicit unsafe override.",
    "title" : "DeliveryLine API",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v1/runner-executions/{rexId}/logs/download" : {
      "get" : {
        "description" : "Returns the runner execution's ALREADY-redacted stdout/stderr as a text/plain attachment (story 3.6 post-hoc redaction is the authoritative guarantee; served verbatim, never re-redacted). Served only over the localhost-only binding to the single local operator; gated by the view_runner_logs allowed-action. The download is recorded as a best-effort audit.logDownloaded event. An unavailable log or gate denial returns 404 RUNNER_EXECUTION_NOT_FOUND.",
        "operationId" : "downloadRunnerLog",
        "parameters" : [
          {
            "description" : "Runner execution public id, e.g. rex_abc123.",
            "example" : "rex_abc123",
            "in" : "path",
            "name" : "rexId",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          },
          {
            "description" : "Actor role for action gating; view_runner_logs is role-agnostic in the runner-execution states, so the default resolves the gate for any operator.",
            "example" : "workflow_owner",
            "in" : "query",
            "name" : "actorRole",
            "required" : false,
            "schema" : {
              "enum" : [
                "product_reviewer",
                "workflow_owner",
                "developer"
              ],
              "type" : "null"
            }
          },
          {
            "in" : "header",
            "name" : "X-Actor-Identity",
            "required" : false,
            "schema" : {
              "type" : "string"
            }
          }
        ],
        "responses" : {
          "200" : {
            "content" : {
              "text/plain" : { }
            },
            "description" : "The redacted runner log as a text/plain attachment."
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetailsResponse"
                }
              }
            },
            "description" : "Malformed runner-execution id (INVALID_ID_PREFIX)."
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetailsResponse"
                }
              }
            },
            "description" : "No such runner execution, no captured/servable log, or view_runner_logs not allowed (RUNNER_EXECUTION_NOT_FOUND)."
          }
        },
        "summary" : "Download a runner execution's redacted log",
        "tags" : [
          "Runner executions"
        ]
      }
    },
...
  },
  "servers" : [
    {
      "description" : "Localhost (loopback-only, Phase 1)",
      "url" : "http://127.0.0.1:8080"
    }
  ],
...
}